Fix browser_list showing incorrect 'FREE mode' error in PRO mode
Fixed variable name typo where _handleBrowserList() was checking
this._mcpConnection (which doesn't exist) instead of this._proxyConnection.

This caused browser_list to incorrectly show "You are currently in FREE mode"
error even when the user was authenticated and connected in PRO mode.

The bug occurred because:
- Status header checks this._isAuthenticated to show "PRO" or "FREE"
- But browser_list checked this._mcpConnection (typo, should be _proxyConnection)
- Since _mcpConnection was undefined, it always triggered the FREE mode error

Changes:
- Line 863: Changed if (!this._mcpConnection) to if (!this._proxyConnection)
- Line 881: Changed this._mcpConnection.sendRequest to this._proxyConnection.sendRequest

This fixes the confusing scenario where users see "PRO v1.9.2" in the status
but get "You are currently in FREE mode" error when trying to list browsers.